Split and migrate network validators tests, remove tests that were moved to nwa-stdlib/pydantic-forms
